Commit working changes to operator composable and component
