+++ apps/web/src/composables/user.ts
@@ -1,7 +1,7 @@
 import { onMounted, onUnmounted, readonly, ref, watch } from 'vue'
 import * as Session from 'supertokens-web-js/recipe/session'
 import { ethers } from 'ethers'
-import { Account, BreakdownAmount, BreakdownString, ContractEventsByAddress, Currency, LoginCredentials, Pool, ProviderString, RegisteredOperator, User, UserWithAccountsAndOperators } from '@casimir/types'
+import { Account, BreakdownAmount, BreakdownString, ContractEventsByAddress, LoginCredentials, UserWithAccountsAndOperators } from '@casimir/types'
 import useContracts from '@/composables/contracts'
 import useEnvironment from '@/composables/environment'
 import useEthers from '@/composables/ethers'
@@ -10,11 +10,10 @@ import useLedger from '@/composables/ledger'
 import usePrice from '@/composables/price'
 import useTrezor from '@/composables/trezor'
 import useWalletConnect from '@/composables/walletConnectV2'
-import { Operator, Scanner } from '@casimir/ssv'
 
 // Test address: 0xd557a5745d4560B24D36A68b52351ffF9c86A212
-const { manager, registry, views } = useContracts()
-const { ethereumUrl, ssvNetworkAddress, ssvNetworkViewsAddress, usersUrl } = useEnvironment()
+const { manager} = useContracts()
+const { ethereumUrl, usersUrl } = useEnvironment()
 const { ethersProviderList, loginWithEthers } = useEthers()
 const { formatNumber } = useFormat()
 const { loginWithLedger } = useLedger()
@@ -24,9 +23,7 @@ const { loginWithWalletConnectV2, initializeWalletConnect, uninitializeWalletCon
 
 const initializeComposable = ref(false)
 const initializedUser = ref(false)
-const nonregisteredOperators = ref<Operator[]>([])
 const provider = new ethers.providers.JsonRpcProvider(ethereumUrl)
-const registeredOperators = ref<Operator[]>([])
 const user = ref<UserWithAccountsAndOperators | undefined>(undefined)
 
 const currentStaked = ref<BreakdownAmount>({
@@ -73,23 +70,6 @@ export default function useUser() {
             return { error, message, data: updatedUser }
         } catch (error: any) {
             throw new Error(error.message || 'Error adding account')
-        }
-    }
-
-    async function addOperator({ address, nodeUrl }: { address: string, nodeUrl: string }) {
-        try {
-            const requestOptions = {
-                method: 'POST',
-                headers: {
-                    'Content-Type': 'application/json'
-                },
-                body: JSON.stringify({ address, nodeUrl })
-            }
-            const response = await fetch(`${usersUrl}/user/add-operator`, requestOptions)
-            const { error, message } = await response.json()
-            return { error, message }
-        } catch (error: any) {
-            throw new Error(error.message || 'Error adding operator')
         }
     }
 
@@ -261,69 +241,12 @@ export default function useUser() {
         }
     }
 
-    async function getUserOperators(): Promise<void> {
-        const userAddresses = (user.value as UserWithAccountsAndOperators).accounts.map((account: Account) => account.address) as string[]
-
-        const scanner = new Scanner({ 
-            ethereumUrl,
-            ssvNetworkAddress,
-            ssvNetworkViewsAddress
-        })
-
-        const ssvOperators: Operator[] = []
-        for (const address of userAddresses) {
-            const userOperators = await scanner.getOperators(address)
-            ssvOperators.push(...userOperators)
-        }
-
-        const casimirOperators: RegisteredOperator[] = []
-        for (const operator of ssvOperators) {
-            const { active, collateral, poolCount, resharing } = await registry.getOperator(operator.id)
-            const registered = active || collateral.gt(0) || poolCount.gt(0) || resharing
-            if (registered) {
-                const pools = await _getPools(operator.id)
-                // TODO: Replace these Public Nodes URLs once we have this working again
-                const operatorStore = {
-                    '654': 'https://nodes.casimir.co/eth/goerli/dkg/1',
-                    '655': 'https://nodes.casimir.co/eth/goerli/dkg/2',
-                    '656': 'https://nodes.casimir.co/eth/goerli/dkg/3',
-                    '657': 'https://nodes.casimir.co/eth/goerli/dkg/4',
-                    '658': 'https://nodes.casimir.co/eth/goerli/dkg/5',
-                    '659': 'https://nodes.casimir.co/eth/goerli/dkg/6',
-                    '660': 'https://nodes.casimir.co/eth/goerli/dkg/7',
-                    '661': 'https://nodes.casimir.co/eth/goerli/dkg/8'
-                }
-                const url = operatorStore[operator.id.toString() as keyof typeof operatorStore]
-                casimirOperators.push({
-                    ...operator,
-                    active,
-                    collateral: ethers.utils.formatEther(collateral),
-                    poolCount: poolCount.toNumber(),
-                    url,
-                    resharing,
-                    pools
-                })
-            }
-        }
-        
-        const nonregOperators = ssvOperators.filter((operator: any) => {
-            const idRegistered = casimirOperators.find((registeredOperator: any) => registeredOperator.id === operator.id)
-            return !idRegistered
-        })
-
-        nonregisteredOperators.value = nonregOperators as Array<Operator>
-        registeredOperators.value = casimirOperators as Array<RegisteredOperator>
-    }
-
     function listenForContractEvents() {
         stopListeningForContractEvents() // Clear old listeners
         try {
             manager.on('StakeDeposited', stakeDepositedListener)
             manager.on('StakeRebalanced', stakeRebalancedListener)
             manager.on('WithdrawalInitiated', withdrawalInitiatedListener)
-            registry.on('OperatorRegistered', getUserOperators)
-            // registry.on('OperatorDeregistered', getUserOperators)
-            // registry.on('DeregistrationRequested', getUserOperators)
         } catch (err) {
             console.log(`There was an error in listenForContractEvents: ${err}`)
         }
@@ -376,7 +299,7 @@ export default function useUser() {
             watch(user, async () => {
                 if (user.value && !initializedUser.value) {
                     initializedUser.value = true
-                    await Promise.all([refreshBreakdown(), getUserOperators()])
+                    await Promise.all([refreshBreakdown()])
                 } else if (!user.value) {
                     uninitializeUser()
                 }
@@ -484,7 +407,7 @@ export default function useUser() {
     }
 
     function uninitializeUser() {
-        // Update user accounts, operators
+        // Update user accounts
         user.value = undefined
         // Reset currentStaked, totalWalletBalance, and stakingRewards
         currentStaked.value = {
@@ -532,37 +455,12 @@ export default function useUser() {
 
     return {
         currentStaked: readonly(currentStaked),
-        nonregisteredOperators: readonly(nonregisteredOperators),
-        registeredOperators: readonly(registeredOperators),
         stakingRewards: readonly(stakingRewards),
         totalWalletBalance: readonly(totalWalletBalance),
         user: readonly(user),
         addAccountToUser,
-        addOperator,
         login,
         logout,
         updateUserAgreement
     }
-}
-
-async function _getPools(operatorId: number): Promise<Pool[]> {
-    const pools: Pool[] = []
-
-    const poolIds = [
-        ...await manager.getPendingPoolIds(),
-        ...await manager.getStakedPoolIds()
-    ]
-
-    for (const poolId of poolIds) {
-        const poolDetails = await views.getPoolDetails(poolId)
-        const pool = {
-            ...poolDetails,
-            operatorIds: poolDetails.operatorIds.map(id => id.toNumber()),
-            reshares: poolDetails.reshares.toNumber()
-        }
-        if (pool.operatorIds.includes(operatorId)) {
-            pools.push(pool)
-        }
-    }
-    return pools
 }
